refactor(toolbar): use HTML and CSS in template literal
Instead of building the toolbar elements imperatively, it is now
declared in HTML and CSS inside of template literals. This makes it much
easier to add layout elements or add styling via CSS.

The toolbar now also uses internal functions by Astro to create the
window element, to synchronize the window placement and to close the
window on outside click. This removes some of the previous hacks and
refactors the window to be more like the first party integrations.