feat(auth): authenticate loader with impersonate token
This allows developers to specify only one impersonate token instead of
superuser credentials with email and password. This makes it much safer
to user the loader since these tokens have a lifetime and can be
invalidated independent of an admins password. Though I still recommend
some kind of technical admin account for this.